@@ -12,9 +12,16 @@
         }
     ],
     "require": {
-        "php": ">=5.2.0"
+        "php": "^5.4 || ^7.0",
+        "ext-bcmath": "*"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^4.8.36 || ^5.6"
     },
     "autoload": {
-        "psr-4": { "": "src/" }
+        "psr-4": { "HappyTypes\\": "src/" }
+    },
+    "autoload-dev": {
+        "psr-4": { "HappyTypes\\Test\\MoneyType\\": "tests/" }
     }
 }
